update notebook viewer
